refactor(auth): import get_current_user from its canonical location
Drop the app.api.v1.auth compat re-export and point every caller at app.lib.auth directly, so
get_current_user has a single canonical home and dependency_overrides cannot split across two import
paths.
