+++ app/api/v1/analytics/routes.py
@@ -14,9 +14,9 @@ from pydantic import ValidationError
 from sqlmodel.ext.asyncio.session import AsyncSession
 
 from app.api.v1.analytics.schemas import EmitEventRequest, EmitEventResponse
-from app.api.v1.auth import get_current_user
 from app.core.analytics.registry import EventRegistry
 from app.lib.analytics import UnknownEventTypeError, ingest_event
+from app.lib.auth import get_current_user
 from app.lib.db import get_analytics_session
 from app.lib.log import get_logger
 from app.models.user import User
