+++ app/core_plugins/chat/routes/attachments.py
@@ -3,7 +3,6 @@ from typing import cast
 from fastapi import APIRouter, Depends, HTTPException, status
 from sqlmodel.ext.asyncio.session import AsyncSession
 
-from app.api.v1.auth import get_current_user
 from app.core_plugins.chat.models import Conversation
 from app.core_plugins.chat.routes.dependencies import get_owned_conversation
 from app.core_plugins.chat.schemas import (
@@ -12,6 +11,7 @@ from app.core_plugins.chat.schemas import (
     ConversationAttachmentResponse,
 )
 from app.core_plugins.chat.service import ChatService, get_chat_service
+from app.lib.auth import get_current_user
 from app.lib.db import get_async_session
 from app.models.user import User
