+++ app/core_plugins/chat/routes/dependencies.py
@@ -4,9 +4,9 @@ from uuid import UUID
 from fastapi import Depends, HTTPException, status
 from sqlmodel.ext.asyncio.session import AsyncSession
 
-from app.api.v1.auth import get_current_user
 from app.core_plugins.chat.models import Conversation
 from app.core_plugins.chat.service import ChatService, get_chat_service
+from app.lib.auth import get_current_user
 from app.lib.db import get_async_session
 from app.models.user import User
