+++ app/core_plugins/googledrive/tests/conftest.py
@@ -16,9 +16,9 @@ import pytest
 from httpx import ASGITransport, AsyncClient
 from sqlmodel.ext.asyncio.session import AsyncSession
 
-from app.api.v1.auth import get_current_user
 from app.core_plugins.googledrive.config import GoogleDriveSettings
 from app.core_plugins.googledrive.models import DriveFile, DriveFolder, DriveOAuthToken
+from app.lib.auth import get_current_user
 from app.main import app
 from app.models.user import User
