+++ tests/api/v1/test_list_user_plugins.py
@@ -4,7 +4,7 @@ from fastapi import FastAPI
 from httpx import ASGITransport, AsyncClient
 from sqlmodel.ext.asyncio.session import AsyncSession
 
-from app.api.v1.auth import get_current_user
+from app.lib.auth import get_current_user
 from app.models import User
 from app.services.plugin import PluginService, get_plugin_service
