+++ tests/api/v1/test_user.py
@@ -12,8 +12,8 @@ from httpx import ASGITransport, AsyncClient
 from sqlalchemy.orm import make_transient
 from sqlmodel.ext.asyncio.session import AsyncSession
 
-from app.api.v1.auth import get_current_user
 from app.core.permissions.models import Role
+from app.lib.auth import get_current_user
 from app.lib.permissions import assign_role
 from app.lib.permissions.scopes import GLOBAL, PermissionScope
 from app.models.user import User
